requirements.txt
google-generativeai>=0.8.0
groq>=0.11.0
huggingface-hub>=0.25.0
python-dotenv>=1.0.0
pyyaml>=6.0
